+++ libs/core/src/infrastructure/adapters/index.ts
@@ -5,3 +5,8 @@ export * from "./GetSmsAdapter"
 
 // Telegram
 export * from "./TelegramAdapter"
+
+// Email
+export * from "./SmtpAdapter"
+export * from "./SendGridAdapter"
+export * from "./MailgunAdapter"